версия 19/requirements.txt
aiohttp>=3.8.0
ecdsa>=0.18.0
base58>=2.1.0
pycryptodome>=3.15.0
Pillow>=9.0.0
requests>=2.28.0
orjson>=3.8.0
uvloop>=0.17.0
async-timeout>=4.0.2
aiodns>=3.0.0
aiofiles>=23.1.0
psutil>=5.9.0
msgpack>=1.0.0